requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
jinja2==3.1.2
torch==2.0.1
transformers==4.30.2
PyGithub==1.58.0
beautifulsoup4==4.12.2
requests==2.31.0
accelerate==0.20.3

